Refactor backend configuration into modular helpers
